Task #4: LLM-generated quiz questions with prefetching, validation, and fallback
Replace hardcoded quiz questions with LLM-generated ones via OpenAI AI Integrations.

Server-side:
- Created server/questionGenerator.ts with world-specific prompts for 5 quiz types
- Added POST /api/generate-questions endpoint with:
  - Session-based authentication (401 for unauthenticated)
  - Per-user rate limiting (10 req/min)
  - worldId whitelist validation (400 for invalid)
  - Count clamping (1-12)
- Uses gpt-5-mini model via Replit AI Integrations

Client-side prefetch store (useQuestionPrefetch.ts):
- Zustand store with prefetchQuestions/getQuestions interface
- Full schema validation: field types, 4 distinct options, correctIndex bounds
- Content-based dedup via question text signatures (avoids ID collisions)
- Invalid questions replaced with random hardcoded fallbacks
- Target count maintained by backfilling from fallback pool
- Only marks succeeded on successful generation (allows retry on failure)
- Seamless fallback to hardcoded questions when LLM unavailable

Prefetch triggers in useGame.tsx:
- Town + town-lesson prefetched on game start
- Ocean, Factory, Psychic prefetched on respective portal entry

Hardcoded questions extracted to separate files:
- townQuestions.ts, factoryQuestions.ts, oceanQuestions.ts, psychicQuestions.ts

All quiz components wired: PracticeQuizUI, FactoryPracticeQuizUI,
OceanPracticeQuizUI, PsychicPracticeQuizUI, GameHUD lesson quiz
