+++ package.json
@@ -68,6 +68,9 @@
     "meshline": "^3.3.1",
     "next-themes": "^0.4.5",
     "ogl": "^1.0.11",
+    "openai": "^6.33.0",
+    "p-limit": "^7.3.0",
+    "p-retry": "^7.1.1",
     "passport": "^0.7.0",
     "passport-local": "^1.0.0",
     "pg": "^8.20.0",
@@ -100,7 +103,7 @@
     "wouter": "^3.3.5",
     "ws": "^8.18.0",
     "zod": "^3.25.76",
-    "zod-validation-error": "^3.4.0",
+    "zod-validation-error": "^3.5.4",
     "zustand": "^5.0.12"
   },
   "devDependencies": {
